Add files via upload
Its a quiz app with registration and login where each username must be unique, if not than a prompt is given username already taken. It consists the use of servlets , jdbc , bootstrap and html, css for connecting to the db and userfriendly UI. It also has a timer for quiz and shows the leader board which is the aggregate of all the quiz a user has given and top 3 users with high scores are shown in the leaderboard. there is also an option of view results where a user can see the score of all the quiz they have given.