Add files via upload
A fun quiz web app built with Java Servlets. Users can register with a unique username (the system prompts if the name is already taken) and log in to play timed quizzes. Each user can view their past results, showing the score of every quiz they’ve attempted. Each time the user takes a quiz they get a different set of questions from the list of questions stored in the DB. There’s also a leaderboard that highlights the top players, showing aggregate high scores across all users, with a podium view for the top three. The UI is styled with Bootstrap for a clean and responsive experience